+++ engine/spaces/native_matching.pl
@@ -606,6 +606,7 @@ clear_native_atoms_stored(Space) :-
     metta_prune_arrow_products(Space),
     metta_capacity_count_cleared(Space),
     retractall(import_life(Space, _, _)),
+    retractall(import_nested_source(Space, _, _)),
     (   SupportModule \== none
     ->  support_graph:support_clear_module(SupportModule)
     ;   true